+++ shelf_modular/test/src/infra/services/route_service_impl_test.dart
@@ -42,4 +42,11 @@ void main() {
       expect(result.isRight, true);
     });
   });
+
+  group('reassemble', () {
+    test('return unit', () async {
+      final result = service.reassemble();
+      expect(result.isRight, true);
+    });
+  });
 }