+++ shelf_modular/test/src/presenter/modular_base_test.dart
@@ -13,6 +13,7 @@ import 'package:shelf_modular/src/domain/usecases/get_arguments.dart';
 import 'package:shelf_modular/src/domain/usecases/get_bind.dart';
 import 'package:shelf_modular/src/domain/usecases/get_route.dart';
 import 'package:shelf_modular/src/domain/usecases/module_ready.dart';
+import 'package:shelf_modular/src/domain/usecases/reassemble_tracker.dart';
 import 'package:shelf_modular/src/domain/usecases/release_scoped_binds.dart';
 import 'package:shelf_modular/src/domain/usecases/report_push.dart';
 import 'package:shelf_modular/src/domain/usecases/start_module.dart';
@@ -48,6 +49,8 @@ class DisposableMock extends Mock implements Disposable {}
 
 class ReportPushMock extends Mock implements ReportPush {}
 
+class ReassembleTrackerMock extends Mock implements ReassembleTracker {}
+
 void main() {
   final disposeBind = DisposeBindMock();
   final getBind = GetBindMock();
@@ -58,6 +61,7 @@ void main() {
   final releaseScopedBinds = ReleaseScopedBindsMock();
   final isModuleReadyImpl = IsModuleReadyImplMock();
   final reportPush = ReportPushMock();
+  final reassembleTracker = ReassembleTrackerMock();
 
   late IModularBase modularBase;
 
@@ -75,7 +79,8 @@ void main() {
         getRoute,
         getArguments,
         releaseScopedBinds,
-        reportPush);
+        reportPush,
+        reassembleTracker);
   });
 
   test('dispose', () {
@@ -113,7 +118,13 @@ void main() {
   test('start (call)', () {
     final module = RouteContextMock();
     when(() => startModule.call(module)).thenReturn(right(unit));
-    final handler = modularBase.call(module: module);
+    final handler =
+        modularBase.call(module: module, middlewares: [MyGuard(true)]);
+
+    when(() => reassembleTracker.call()).thenReturn(right(unit));
+
+    modularBase.reassemble();
+
     verify(() => startModule.call(module)).called(1);
     expect(handler, isA<FutureOr<Response> Function(Request request)>());
     expect(() => modularBase.start(module: module),
@@ -139,14 +150,23 @@ void main() {
     expect(result.statusCode, 200);
   });
 
-  test('disposeBindFunction', () {
-    final disposableMock = DisposableMock();
-    (modularBase as ModularBase).disposeBindFunction(disposableMock);
-    verify(() => disposableMock.dispose()).called(1);
-  });
-
   test('handler with error', () async {
     final request = RequestMock();
+    final route = RouteMock();
+
+    when(() => request.method).thenReturn('GET');
+    when(() => request.url).thenReturn(Uri.parse(''));
+    when(() => route.handler).thenReturn((String v) {});
+    when(() => route.middlewares).thenReturn([]);
+
+    when(() => getArguments.call()).thenReturn(right(ModularArguments.empty()));
+
+    when(() => request.method).thenReturn('GET');
+    when(() => request.url).thenReturn(Uri.parse(''));
+    when(() => releaseScopedBinds.call()).thenReturn(right(unit));
+    when(() => getRoute.call(any())).thenThrow(Error());
+
+    when(() => reportPush.call(route)).thenReturn(right(unit));
     when(() => releaseScopedBinds.call()).thenReturn(right(unit));
 
     final result = await (modularBase as ModularBase).handler(request);
@@ -172,6 +192,8 @@ void main() {
     when(() => releaseScopedBinds.call()).thenReturn(right(unit));
     when(() => getRoute.call(any())).thenAnswer((_) async => right(route));
 
+    when(() => reportPush.call(route)).thenReturn(right(unit));
+
     final result = await (modularBase as ModularBase).handler(request);
     expect(result.statusCode, 404);
   });
@@ -182,12 +204,14 @@ void main() {
 
     when(() => request.method).thenReturn('GET');
     when(() => request.url).thenReturn(Uri.parse(''));
-    when(() => route.handler).thenReturn(() {});
+    when(() => route.handler).thenReturn((String v) {});
     when(() => route.middlewares).thenReturn([]);
 
     when(() => releaseScopedBinds.call()).thenReturn(right(unit));
     when(() => getArguments.call()).thenReturn(right(ModularArguments.empty()));
     when(() => getRoute.call(any())).thenAnswer((_) async => right(route));
+
+    when(() => reportPush.call(route)).thenReturn(right(unit));
 
     final result = await (modularBase as ModularBase).handler(request);
     expect(result.statusCode, 500);
@@ -207,6 +231,8 @@ void main() {
     when(() => getRoute.call(any()))
         .thenAnswer((_) async => left(RouteNotFoundException('')));
 
+    when(() => reportPush.call(route)).thenReturn(right(unit));
+
     final result = await (modularBase as ModularBase).handler(request);
     expect(result.statusCode, 404);
   });
@@ -224,6 +250,28 @@ void main() {
     when(() => getArguments.call()).thenReturn(right(ModularArguments.empty()));
     when(() => getRoute.call(any()))
         .thenAnswer((_) async => left(ModuleStartedException('')));
+
+    when(() => reportPush.call(route)).thenReturn(right(unit));
+
+    final result = await (modularBase as ModularBase).handler(request);
+    expect(result.statusCode, 500);
+  });
+
+  test('handler error \'Handler not correct\'', () async {
+    final request = RequestMock();
+    final route = RouteMock();
+
+    when(() => request.method).thenReturn('GET');
+    when(() => request.url).thenReturn(Uri.parse(''));
+    when(() => route.handler).thenReturn(() {});
+    when(() => route.middlewares).thenReturn([]);
+
+    when(() => releaseScopedBinds.call()).thenReturn(right(unit));
+    when(() => getArguments.call()).thenReturn(right(ModularArguments.empty()));
+    when(() => getRoute.call(any()))
+        .thenAnswer((_) async => left(ModuleStartedException('')));
+
+    when(() => reportPush.call(route)).thenReturn(right(unit));
 
     final result = await (modularBase as ModularBase).handler(request);
     expect(result.statusCode, 500);
@@ -244,10 +292,13 @@ void main() {
     when(() => getArguments.call()).thenReturn(right(ModularArguments.empty()));
     when(() => getRoute.call(any())).thenAnswer((_) async => right(route));
 
+    when(() => reportPush.call(route)).thenReturn(right(unit));
+
     final result = await (modularBase as ModularBase).handler(request);
     expect(result.statusCode, 403);
 
     expect(MyGuard(true).pre(route), route);
+    expect(MyGuard(true).pos(route, request), route);
   });
 
   test('tryJsonDecode isMultipart false', () async {